# Connection: Remove excessive verbose logging