--- a/lib/connection.js
+++ b/lib/connection.js
@@ -434,7 +434,6 @@ Connection.prototype.sendStream = function (request, options, callback) {
  * @private
  */
 Connection.prototype._write = function (operation, streamId) {
-  this.log('verbose', 'Sending stream #' + streamId);
   operation.streamId = streamId;
   var self = this;
   this.writeQueue.push(operation, function writeCallback (err) {
@@ -520,7 +519,6 @@ Connection.prototype.freeStreamId = function(header) {
     this.emit('drain');
   }
   this._writeNext();
-  this.log('verbose', 'Done receiving frame #' + streamId);
 };
 
 Connection.prototype._writeNext = function () {
@@ -589,7 +587,6 @@ Connection.prototype.handleRow = function (header, row, meta, rowLength, flags) 
   if (!operation) {
     return this.log('error', 'The server replied with a wrong streamId #' + streamId);
   }
-  this.log('verbose', 'Received streaming frame #' + streamId);
   operation.setResultRow(row, meta, rowLength, flags);
 };