I've been benchmarking possible ways to limit the overhead of verbose message generation and event emission on the Connection.

The final solution is simple enough, just remove the excess (instead of "sending" & "sent", leave only "sent" logs; similar for "received" and "done"). I've tested using delayed string evaluation, a wrapper with references to the actual values, but there isn't any significant difference (again: string manipulation seems to be highly optimized on v8).

Removing these log message (even if they are not consumed) causes an increase in throughput of 7% under Node.js 6 and 3.6% under Node.js 8:

Node.js 6
- Baseline: 17176.23 req/sec
- This pr: 18381.51 req/sec (+7%)

Node.js 8
- Baseline: 33742.75 req/sec
- This pr: 34966.87 req/sec (+3.6%)

It uses the throughput test, with a single node cluster with 320 max outstanding requests.

_oh my, look at those Node.js 8 numbers!_ 😃